requirements.txt
requests
beautifulsoup4
pandas
google-genai
python-telegram-bot